Move scroll position struct to view
The scrolling functionality is directly related to the view, so move the
struct into the view module.
