+++ src/list/mod.rs
@@ -22,8 +22,8 @@ use crate::process::handle_input_result::{HandleInputResult, HandleInputResultBu
 use crate::process::process_module::ProcessModule;
 use crate::process::process_result::ProcessResult;
 use crate::process::state::State;
-use crate::scroll::scroll_position::ScrollPosition;
 use crate::view::line_segment::LineSegment;
+use crate::view::scroll_position::ScrollPosition;
 use crate::view::view_line::ViewLine;
 use crate::view::View;
 use std::cmp;
