--- src/scroll/mod.rs
@@ -1,1 +1,0 @@
-pub(crate) mod scroll_position;
